% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{cv.intdag.coe.penalized}
\alias{cv.intdag.coe.penalized}
\title{Structure Learning Implementation}
\usage{
cv.intdag.coe.penalized(X, Y, Pi, Phi, Piv, tau.list, gamma.list, n.fold)
}
\arguments{
\item{X}{A n*q matrix for intervention variables.}

\item{Y}{A n*p matrix for primary variables.}

\item{Pi}{A matrix representing non-zero elements' indices for U.}

\item{Phi}{A matrix representing non-zero elements' indices for W.}

\item{Piv}{A matrix representing non-zero elements' indices for candidate IVs.}

\item{tau.list}{A set of taus to be selected for TLP.}

\item{gamma.list}{A set pf gammas to be selected for TLP.}

\item{n.fold}{Number of folds in cross-validation.}
}
\value{
A list consists of the estimates of U,W and covariance matrix as well as selected taus and gammas.
}
\description{
This function implements coefficient estimation, however, the parameter of interests are also penalized.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
