% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{cv.intdag.pmle.diff.aic}
\alias{cv.intdag.pmle.diff.aic}
\title{Matrix V Estimation Function with Cross-Validation}
\usage{
cv.intdag.pmle.diff.aic(X, Y, tau.list, gamma.list, n.fold)
}
\arguments{
\item{X}{A n*q matrix for intervention variables.}

\item{Y}{A n*p matrix for primary variables.}

\item{tau.list}{A list of parameters to be selected for TLP in each regression model where Y[,i] is the response with X the predictor matrix.}

\item{gamma.list}{A list of parameters to be selected for TLP in each regression model where Y[,i] is the response with X the predictor matrix.}

\item{n.fold}{Number of folds to be used.}
}
\value{
A list consists of an estimate of matrix V, taus and gammas selected
}
\description{
This function estimates matrix V with cross-validation.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
