% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlpreg.R
\name{cv.tlpreg0}
\alias{cv.tlpreg0}
\title{Cross-valdiated Function of Truncated L1-Penalty Sparse Regression}
\usage{
cv.tlpreg0(
  y,
  X,
  tau.list,
  gamma.list,
  n.fold,
  b.init = rep(0, ncol(X)),
  pen.fac = rep(1, ncol(X)),
  tol = 1e-05,
  dc.maxit = 20,
  cd.maxit = 10000
)
}
\arguments{
\item{y}{A vector of responses.}

\item{X}{A n*p matrix of predictor variables.}

\item{tau.list}{Parameter to be selected for TLP.}

\item{gamma.list}{Parameter to be selected for TLP.}

\item{n.fold}{Folds for cross-validation.}

\item{b.init}{Initial values for coefficient estimation.}

\item{pen.fac}{Coefficients to be penalized.}

\item{tol}{Stopping criterion.}

\item{dc.maxit}{Maximum number of DC repetitions.}

\item{cd.maxit}{Maximum number of corrdinate descents.}
}
\value{
An estimate of coefficients.
}
\description{
This function conducts cross-validate to selection tunning parameters for truncated L1 penalized sparse regression
}
\examples{
\dontrun{
to be added.
}
}
