% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{intdag.2lr}
\alias{intdag.2lr}
\title{Statistic Computation}
\usage{
intdag.2lr(X, Y, U0, W0, U1, W1, S)
}
\arguments{
\item{X}{A n*q matrix representing interventions.}

\item{Y}{A n*p matrix representing primary variables.}

\item{U0}{Matrix of Indices for non-zero elements of U under the null hypothesis.}

\item{W0}{Matrix of Indices for non-zero elements of W under the null hypothesis.}

\item{U1}{Matrix of Indices for non-zero elements of U under the alternative hypothesis.}

\item{W1}{Matrix of Indices for non-zero elements of W under the alternative hypothesis.}

\item{S}{Precision matrix.}
}
\value{
A list consits of estimates of U,W under the null and alternative hypothesis and the statistic.
}
\description{
This function computes 2Lr on the whole graph.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
