% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{intdag.localization.ver2}
\alias{intdag.localization.ver2}
\title{Statistic Computation}
\usage{
intdag.localization.ver2(X, Y, U.diff, U.hat, W.hat, Pi, Phi, Sigma)
}
\arguments{
\item{X}{A n*q matrix representing interventions.}

\item{Y}{A n*p matrix representing primary variables.}

\item{U.diff}{Edges to be tested in the hypothesis testing.}

\item{U.hat}{U or estimates of U.}

\item{W.hat}{W or estimates of W.}

\item{Pi}{Ancestral relationships.}

\item{Phi}{Interventional relationships.}

\item{Sigma}{Covariance matrix.}
}
\value{
A list consits of related interventions X, related primary variable Y, and related estimates of U,W,Pi,Phi and Sigma.
}
\description{
This function computes the local graph with nodes involved in the tests and their ancestors.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
