% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtlp.R
\name{precision_refit}
\alias{precision_refit}
\title{Precision Matrix Estimation}
\usage{
precision_refit(Sigma, S, max.it, tol)
}
\arguments{
\item{Sigma}{Covariance matrix.}

\item{S}{Support of the precision matrix.}

\item{max.it}{Maximum number of iterations allowed.}

\item{tol}{Tolerance of stopping criterion.}
}
\value{
An estimate of precision matrix
}
\description{
This function refit the precision matrix using BCD algorithm with details given in Appendix.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
