% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlpreg.R
\name{tlpreg0}
\alias{tlpreg0}
\title{Truncated L1-Penalty Sparse Regression}
\usage{
tlpreg0(
  y,
  X,
  b.init = rep(0, ncol(X)),
  tau = 0.01,
  gamma = 0.5,
  pen.fac = rep(1, ncol(X)),
  tol = 1e-05,
  dc.maxit = 20,
  cd.maxit = 10000
)
}
\arguments{
\item{y}{A vector of responses.}

\item{X}{A n*p matrix of predictor variables.}

\item{b.init}{Initial values for coefficient estimation.}

\item{tau}{Parameter used for TLP.}

\item{gamma}{Parameter used for TLP.}

\item{pen.fac}{Coefficients to be penalized.}

\item{tol}{Stopping criterion.}

\item{dc.maxit}{Maximum number of DC repetitions.}

\item{cd.maxit}{Maximum number of corrdinate descents.}
}
\value{
An estimate of coefficients.
}
\description{
This function conducts truncated L1 penalized sparse regression
}
\examples{
\dontrun{
to be added.
}
}
