% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{topological_order}
\alias{topological_order}
\title{Peeling Algorithm Implementation}
\usage{
topological_order(v)
}
\arguments{
\item{v}{A q*p matrix encoding relationships among Y and X.}
}
\value{
A list consists of the estimates of ancestral relationships, interventional relationships and candidate-iv relationships.
}
\description{
This function conducts the peeling algorithm.
}
\examples{
\dontrun{
to be added.
}
}
\references{
Discovery and inference of a causal network with hidden confounding.
}
